deps update
This was part of a debug phase. I mostly tried to keep up with the
current chestnut template. Some of the old chestnut deps are gone now,
like figwheel and weasel. Not sure what to think about that, but
everything seems to work at this point. Mostly glad to have cljs and om
in newer forms.